yourstoryHack/gungun/requirements.txt
Django==3.2.4
django-modeltranslation==0.17.1
scikit-learn==0.24.2
